refactor: modularize CLI tests
Split the various subcommand CLI tests into their own files for better
modularization. While maintaining the common imports and functionality
like the HELLO_WORLD_PROGRAM in a separate module with no tests.
